Rename m4_vector_wave to m4_ewt
Refactor experiment namespace from openwave/xperiments/m4_vector_wave to openwave/xperiments/m4_ewt. Renamed/moved the experiment package and its submodules/data/plots/xparameters, updated import paths (notably in _launcher.py and several xparameters files), and fixed cross-references in documentation (CLAUDE.md, MODELS.md, and an M5 roadmap doc) to the new m4_ewt locations and wording. This is a straightforward rename/rebrand of the M4 experiment to the EWT variant and updates related docs and module imports accordingly.
