+++ openwave/xperiments/m4_ewt/xparameters/formation10p.py
@@ -19,7 +19,7 @@ Each K places WCs at λ/2 spacing in the simplest possible geometry:
 """
 
 from openwave.common import constants
-from openwave.xperiments.m4_vector_wave.xparameters.formation02 import generate_K_positions
+from openwave.xperiments.m4_ewt.xparameters.formation02 import generate_K_positions
 
 UNIVERSE_EDGE = 1e-15  # m, universe edge length in meters
 TARGET_VOXELS = 100_000_000  # Target voxel count (impacts performance)
